Tier0 64-bit fixes

+++ src/public/tier0/threadtools.h
@@ -65,7 +65,7 @@ const unsigned TT_INFINITE = 0xffffffff;
 
 #endif // NO_THREAD_LOCAL
 
-typedef unsigned long ThreadId_t;
+typedef unsigned int ThreadId_t;
 
 //-----------------------------------------------------------------------------
 //
@@ -1306,7 +1306,7 @@ private:
 		bool *        pfInitSuccess;
 	};
 
-	static unsigned __stdcall ThreadProc( void * pv );
+	static unsigned int __stdcall ThreadProc( void * pv );
 
 	// make copy constructor and assignment operator inaccessible
 	CThread( const CThread & );
